+++ api/src/types/planets/earth/calendar/traits/day_of_week/xavetar.rs
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Stanislav Mikhailov (xavetar)
+ * Copyright 2024 Stanislav Mikhailov (xavetar)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -47,7 +47,7 @@ use crate::types::{
 
 pub trait Xavetar {
     fn week_day(&self) -> Week;
-    fn from(view: CalendarView, year: u128, month: u8, day: u8) -> Week;
+    fn from(view: CalendarView, year: u64, month: u8, day: u8) -> Week;
 }
 
 impl Xavetar for Date {
@@ -55,7 +55,7 @@ impl Xavetar for Date {
         return <Date as Xavetar>::from(self.view.clone(), self.year, self.month, self.day);
     }
 
-    fn from(view: CalendarView, year: u128, month: u8, day: u8) -> Week {
+    fn from(view: CalendarView, year: u64, month: u8, day: u8) -> Week {
         let (BASE_YEAR_SHIFTS, LEAP_YEAR_SHIFTS): (&[u8; MONTHS_IN_YEAR as usize], &[u8; MONTHS_IN_YEAR as usize]) = match view {
             CalendarView::Solar => (&SOLAR_BASE_XAVETAR, &SOLAR_LEAP_XAVETAR),
             CalendarView::Julian => (&JULIAN_BASE_XAVETAR, &JULIAN_LEAP_XAVETAR),
@@ -63,12 +63,12 @@ impl Xavetar for Date {
             _ => panic!("[ERROR]: Unknown CalendarView for Julian or Gregorian calendar (Xavetar).")
         };
 
-        let last_year: u128 = year - 1_u128;
+        let last_year: u64 = year - 1_u64;
 
         if !is_leap_year(view, year) {
-            return Week::from(((last_year + sum_leap_years(view, last_year) + BASE_YEAR_SHIFTS[(month - 1_u8) as usize] as u128 + day as u128) % REPEAT_WEAK_DAY_CYCLE as u128) as u8);
+            return Week::from(((last_year + sum_leap_years(view, last_year) + BASE_YEAR_SHIFTS[(month - 1_u8) as usize] as u64 + day as u64) % REPEAT_WEAK_DAY_CYCLE as u64) as u8);
         } else {
-            return Week::from(((last_year + sum_leap_years(view, last_year) + LEAP_YEAR_SHIFTS[(month - 1_u8) as usize] as u128 + day as u128) % REPEAT_WEAK_DAY_CYCLE as u128) as u8);
+            return Week::from(((last_year + sum_leap_years(view, last_year) + LEAP_YEAR_SHIFTS[(month - 1_u8) as usize] as u64 + day as u64) % REPEAT_WEAK_DAY_CYCLE as u64) as u8);
         }
     }
 }
